+++ CHANGES.md
@@ -11,9 +11,10 @@ E.g. to retrieve the changes made between release 3.4 and release 3.5, type::
 The contributors lists are in no order and apologies to those accidentally not
 mentioned. If we missed you, please let us know!
 
-CHANGES  4.4.3   ??/06/2025
+CHANGES  4.4.3   23/07/2025
 ---------------------------
     * start implmenting VerticalLineChart
+    * use usedforsecurity=False in most md5 calls; patch from Martin Renters
 
 CHANGES  4.4.2   18/06/2025
 ---------------------------